Add doc `Form::class`.
