+++ docs/tag/Form.md
@@ -1,23 +1,21 @@
-# Footer
+# Form
 
-Represents a footer for its nearest ancestor sectioning content or sectioning root element.
-
-A `footer` typically contains information about the author of the section, copyright data or links to related documents.
+The `<form>` `HTML` element represents a document section containing interactive controls for submitting information.
 
 ## Basic Usage
 
-Instantiate the `Footer` class using `Footer::widget()`.
+Instantiate the `Form` class using `Form::widget()`.
 
 ```php
-$footer = Footer::widget();
+$form = Form::widget();
 ```
 
 Or, block style instantiation.
 
 ```php
-<?= Footer::begin() ?>
-    // ... content to be wrapped by `footer` element
-<?= Footer::end() ?>
+<?= Form::begin() ?>
+    // ... content to be wrapped by `form` element
+<?= Form::end() ?>
 ```
 
 ## Setting Attributes
@@ -26,29 +24,29 @@ Use the provided methods to set specific attributes for the a element.
 
 ```php
 // setting href attribute
-$footer->class('container');
+$form->class('container');
 ```
 
 Or, use the `attributes` method to set multiple attributes at once.
 
 ```php
-$footer->attributes(['class' => 'container', 'style' => 'background-color: #eee;']);
+$form->attributes(['class' => 'container', 'style' => 'background-color: #eee;']);
 ```
 
 ## Adding Content
 
-If you want to include content within the `footer` tag, use the `content` method.
+If you want to include content within the `form` tag, use the `content` method.
 
 ```php
-$footer->content('My content');
+$form->content('My content');
 ```
 
 Or, use `begin()` and `end()` methods to wrap content.
 
 ```php
-<?= Footer::begin() ?>
+<?= Form::begin() ?>
     My content
-<?= Footer::end() ?>
+<?= Form::end() ?>
 ```
 
 ## Rendering
@@ -56,19 +54,19 @@ Or, use `begin()` and `end()` methods to wrap content.
 Generate the `HTML` output using the `render` method, for simple instantiation. 
 
 ```php
-$html = $footer->render();
+$html = $form->render();
 ```
 
 For block style instantiation, use the `end()` method, which returns the `HTML` output.
 
 ```php
-$html = Footer::end();
+$html = Form::end();
 ```
 
 Or, use the magic `__toString` method.
 
 ```php
-$html = (string) $footer;
+$html = (string) $form;
 ```
 
 ## Common Use Cases
@@ -77,43 +75,58 @@ Below are examples of common use cases:
 
 ```php
 // adding multiple attributes
-$footer->class('external')->content('My content');
+$form->class('external')->content('My content');
 
 // using data attributes
-$footer->dataAttributes(['analytics' => 'trackClick']);
+$form->dataAttributes(['analytics' => 'trackClick']);
+
+// form with content
+Form::widget()
+    ->action('/foo')
+    ->content('value', PHP_EOL, Span::widget()->content('value'))
+    ->csrf('csrf-token')
+    ->method('POST')
+    ->render()
 ```
 
-Explore additional methods for setting various attributes such as `lang`, `name`, `style`, `title`, etc.
+Explore additional methods for setting various attributes such as `action`, `method`, `name`, `style`, `title`, etc.
 
 ## Attributes
 
-Refer to the [Attribute Tests](https://github.com/php-forge/html/blob/main/tests/Footer/AttributeTest.php) for
+Refer to the [Attribute Tests](https://github.com/php-forge/html/blob/main/tests/Form/AttributeTest.php) for
 comprehensive examples.
 
 The following methods are available for setting attributes:
 
 | Method            | Description                                                                                      |
 | ----------------- | ------------------------------------------------------------------------------------------------ |
+| `accept()`        | Set the `accept` attribute.                                                                      |
+| `action()`        | Set the `action` attribute.                                                                      |
 | `attributes()`    | Set multiple `attributes` at once.                                                               |
+| `autocomplete()`  | Set the `autocomplete` attribute.                                                                |
 | `class()`         | Set the `class` attribute.                                                                       |
-| `content()`       | Set the `content` within the `footer` element.                                                   |
+| `content()`       | Set the `content` within the `form` element.                                                     |
 | `dataAttributes()`| Set multiple `data-attributes` at once.                                                          |
+| `enctype()`       | Set the `enctype` attribute.                                                                     |
 | `id()`            | Set the `id` attribute.                                                                          |
 | `lang()`          | Set the `lang` attribute.                                                                        |
+| `method()`        | Set the `method` attribute.                                                                      |
 | `name()`          | Set the `name` attribute.                                                                        |
+| `novalidate()`    | Set the `novalidate` attribute.                                                                  |
+| `rel()`           | Set the `rel` attribute.                                                                         |
 | `style()`         | Set the `style` attribute.                                                                       |
 | `title()`         | Set the `title` attribute.                                                                       |
 
 ## Custom methods
 
-Refer to the [Custom Methods Tests](https://github.com/php-forge/html/blob/main/tests/Footer/CustomMethodTest.php) for
+Refer to the [Custom Methods Tests](https://github.com/php-forge/html/blob/main/tests/Form/CustomMethodTest.php) for
 comprehensive examples.
 
 The following methods are available for customizing the `HTML` output:
 
 | Method    | Description                                                                                              |
 | --------- | -------------------------------------------------------------------------------------------------------- |
-| `begin() `| Start the `footer` element.                                                                              |
-| `end()`   | End the `footer` element, and generate the `HTML` output.                                                |
+| `begin() `| Start the `form` element.                                                                                |
+| `end()`   | End the `form` element, and generate the `HTML` output.                                                  |
 | `render()`| Generates the `HTML` output.                                                                             |
 | `widget()`| Instantiates the `Body::class`.                                                                          |
